requirements.txt
pandas==2.2.3
numpy==2.2.4
matplotlib==3.10.1
seaborn==0.13.2
scikit-learn==1.6.1
xgboost==2.1.4
lightgbm==4.6.0
joblib==1.4.2
pyyaml==6.0.2
jupyter==1.1.1
ipykernel==6.29.5